fix(ci): python-full-coverage — split oversized test_workspace_services_compose.py under line limit
The maintainability guardrail test_first_party_code_files_stay_under_line_limit
failed because test_workspace_services_compose.py grew to 2126 lines (>1500 limit)
across many PR-#751 hosted-secret-safety fix commits. Split it into a
test_workspace_services_compose_parts/ package following the established
parts-directory pattern (e.g. test_executor_error_paths_parts): part_001 holds
the shared fixtures/helpers and the first half of tests; part_002 imports them
and carries the remaining tests. No test logic changed — only relocated.
